+++ src/libstd/std.rc
@@ -19,7 +19,10 @@ export test, tempfile, serialization;
 // General io and system-services modules
 
 mod net;
+
+// libuv modules
 mod uv;
+mod uv_ll;
 
 
 // Utility modules
